Download Foundry binaries directly from download.farcaster.xyz
Foundry doesn't keep around old artifacts, so our previous pinning
approach had the unintended effect of breaking the build once the
artifact was removed.

Since we are pinning to avoid unexpected breakages in later versions of
Foundry's toolchain, pin to a build artifact that is known to work
(downloaded from GitHub Sept 3, 2023 UTC).

While here, add a test to ensure the image works.