feat(FnameResolver): Add support for ENS name resolution
This PR adds support for the ENS name resolution function to the FnameResolver contract.

Previously, the resolver only supported the addr(bytes32) function, but now it also
supports the name(bytes32) function, which allows reverse resolution from addresses
to names. This makes the resolver more compliant with ENS standards and improves
its functionality.

The implementation:
1. Adds a new INameQuery interface with the name function
2. Updates the resolve function to accept name function calls
3. Modifies resolveWithProof to return different data based on the original function call
4. Adds tests to verify the new functionality

Fixes the "ResolverFunctionNotSupported" error that was previously thrown for all
non-addr function calls, as indicated in the code comment.